refactor: prefetch to preload and extract as utils
