--- packages/preset-umi/templates/routePrefetch/prefetchRouteFilesScp.js
@@ -1,1 +1,0 @@
-!function(){"use strict";var t=function(t,c,e){var n,a=document,i=a.head,r=a.createElement.bind(a),o=e.publicPath,l=c.r[t]||(null===(n=Object.entries(c.r).find((function(c){var e=c[0];return new RegExp("^".concat(e.replace(/\/:[^/]+/g,"/[^/]+").replace("/*","/.+"),"$")).test(t)})))||void 0===n?void 0:n[1]);null==l||l.forEach((function(t){var e,n=c.f[t][1],a=c.f[t][0],l=a.split(".").pop();if("js"===l)(e=r("script")).src="".concat(o).concat(a),e.async=!0;else{if("css"!==l)return;(e=r("link")).href="".concat(o).concat(a),e.rel="preload",e.as="style"}e.setAttribute("data-".concat(c.b),"".concat(c.p,":").concat(n)),i.appendChild(e)}))},c="{{basename}}",e=location.pathname,n=e.startsWith(c)&&decodeURI("/".concat(e.slice(c.length)));n&&t(n,"{{routeChunkFilesMap}}",{publicPath:"{{publicPath}}"})}();
